feat(team): team facade tests

+++ GamingManagement-service/src/main/java/cz/fi/muni/pa165/yellow_yak/facade/TeamFacadeImpl.java
@@ -28,24 +28,36 @@ public class TeamFacadeImpl implements TeamFacade {
 
     @Override
     public TeamDTO create(String name) {
+        if (name == null) {
+            throw new IllegalArgumentException("name cannot be null");
+        }
         log.info("create team, name = {}", name);
         return beanMappingService.mapTo(teamService.create(name), TeamDTO.class);
     }
 
     @Override
     public void remove(Long teamId) {
+        if (teamId == null) {
+            throw new IllegalArgumentException("teamId cannot be null");
+        }
         log.info("removing team, id = {}", teamId);
         teamService.remove(teamId);
     }
 
     @Override
     public TeamDTO findById(Long teamId) {
+        if (teamId == null) {
+            throw new IllegalArgumentException("teamId cannot be null");
+        }
         log.info("finding team by ID, id = {}", teamId);
         return beanMappingService.mapTo(teamService.findById(teamId), TeamDTO.class);
     }
 
     @Override
     public List<TeamDTO> findByName(String name) {
+        if (name == null) {
+            throw new IllegalArgumentException("name cannot be null");
+        }
         log.info("finding teams by name, name = {}", name);
         return beanMappingService.mapTo(teamService.findByName(name), TeamDTO.class);
     }
